% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRpackagedata_to_sas7bdatlibrary.R
\name{exportRpackagedata_to_sas7bdatlibrary}
\alias{exportRpackagedata_to_sas7bdatlibrary}
\title{export all tables from an installed R data package to sas7bdat data files in a given library}
\usage{
exportRpackagedata_to_sas7bdatlibrary(package, path_to_export_to)
}
\arguments{
\item{package}{package to be transformed}

\item{SAS_library_path}{path to export the data}
}
\value{
nothing
}
\description{
export all tables from an installed R data package to sas7bdat data files in a given library
}
\examples{
exportRpackagedata_to_sas7bdatlibrary("datasets")
}
