% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotmissinggenericfunction.R
\name{ggplot_missing}
\alias{ggplot_missing}
\title{Create missing chart}
\usage{
ggplot_missing(x, reordonne = FALSE)
}
\arguments{
\item{x}{a dataframe}

\item{reordonne}{a boolean}
}
\value{
a ggplot graph
}
\description{
Create missing chart
}
\examples{
library(reshape2)
library(ggplot2)
library(plyr)
library(magrittr)
X=cars
for(i in 1:40){
  X[sample(1:50,1,replace=TRUE),sample(1:2,1,replace=TRUE)]<-NA}
ggplot_missing(X,reordonne=TRUE)
ggplot_missing(X,reordonne=FALSE)createallautomaticRMD(schema="SDP")
}
