% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotmissinggenericfunction.R
\name{ggplot_missing2}
\alias{ggplot_missing2}
\title{Create missing chart}
\usage{
ggplot_missing2(X, reordonne = TRUE, keep = NULL)
}
\arguments{
\item{X}{a dataframe}

\item{reordonne}{a boolean}

\item{keep}{a boolean}
}
\value{
a ggplot graph
}
\description{
Create missing chart
}
\examples{
library(reshape2)
library(ggplot2)
library(plyr)
X=cars
X$year=sample(2012:2017,nrow(cars),replace=TRUE)
for(i in 1:40){
 X[sample(1:50,1,replace=TRUE),sample(1:2,1,replace=TRUE)]<-NA}
ggplot_missing2(X,keep="year")
}
