% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_report.R
\name{missing.summary}
\alias{missing.summary}
\title{tablename="TRAIT_LKP"
automaticdataf(tablename)}
\usage{
missing.summary(X, info2 = NULL)
}
\description{
tablename="TRAIT_LKP"
automaticdataf(tablename)
}
