% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{StudyDataTools-package}
\alias{StudyDataTools}
\alias{StudyDataTools-package}
\title{StudyDataTools: X}
\description{
Data
}
\details{
Some non confidential functions developped for a synthetic data project
}
\keyword{internal}
